Save routing matrix without .pyc suffix.
For some unknown reason, when using marshal.dump(), if the output file has a
name with a .pyc suffix, the file doesn't get created and the matrix isn't
saved to disk. No error is raised, though. I suppose I should figure out
why, but I don't think it really matters, and the file format of the matrix
is an implementation detail anyway.