Don't allow adjacency matrix to be created on the fly
I think the original idea behind this was to allow the matrix to be
dynamically built on the first request after a data update. Because
building the graph can take a while (and use a lot of memory), this
is probably a bad idea.

The deployment process should ensure the matrix is built up front,
before restarting the app. It would also be good if the matrix could
be built offline and uploaded to the production server.