Use joinedload to speed up Edge query in Route service
In service.Route._makeDirectionsForPath(), the Edge query would fetch the edge
rows but not the rows for the related attributes (street name and nodes) that
are also used in building the directions list. This would result in 3*N
additional database queries, three for each edge.

Using joinedload with street_name, node_f, and node_t reduces this down to
a single query, which improves performance noticeably in testing.

There is another opportunity for reducing the number of queries here: the
_getDifferentStreetNameFromNode method may be called up to N times from
_makeDirectionsFromPath. Each call to _getDifferentStreetNameFromNode results
in two additional database queries to get the edges for a node. It should be
possible to reduce this down to a single query as well. This will be more
complicated though because we don't know up front which nodes we will need to
grab the edges for (only the ones at the start of a stretch).