Fetch all toward nodes up front when making directions list
This cuts the time it takes to run the route tests roughly in half.

Previously, the _getDifferentStreetNameFromNode method would be called
up to N times from _makeDirectionsFromPath. Each call to
_getDifferentStreetNameFromNode would result in two additional database
queries to get the edges for a node.

This has been reduced down to a single query by gathering the args
needed to get the toward streets and then getting them all as a batch.

Note: It's kind of hilarious that these toward values aren't even being
used in the UI. Oh well. At least the overhead now of fetching them is
almost neglible, and I may want to add them back to the UI. I'll have to
think about that...