Simplify how DB URL is created & don't require .pw
model.db.getConnectionUri() was renamed to make_url(). It's now a simple
wrapper around SQLAlchemy's URI class. The .pw file isn't required any
more and won't be used if it exists.

TODO: The whole DB connection setup stuff is a mess and needs to fixed
so that connection options are (or at least *can* be) read from config.