Don't create encoded polyline for routes
This removes the `linestring_encoded` and `linestring_encoded_levels`
attributes from the `Route` model.

This was originally for use with Google Maps way back when, and it made
a big difference with long lines. Currently, it's not used in the front
end, and I'm not sure it's necessary any more for performance.