Correctly distribute coordinates when splitting a street
Previously, the LAST coordinate in the street's line geometry was being
set as the FIRST coordinate of the second segment, but it should be the
LAST coordinate of the second segment. This caused visually odd results
when displaying routes, but it wasn't always noticeable, esp. for routes
with only right-angle turns.