Save routing graph (AKA matrix) in database
Previously, it was saved to disk, but that meant deciding where to save
it and figuring out how to discover and load it into memory, which was
particularly problematic in production.

Putting the graph in the database removes those issues. In my personal
dev setup, it also means I don't have to worry about inadvertently
backing up the matrix file any more.