[WIP] Add Mapbox address/POI lookup
This replaces `LookupService.match_adress()` and `.match_poi()`, both of
which were no-ops, with `.match_via_mapbox()`.

I started on this a while back and made some decent progress, but it
doesn't fully work yet... and I don't recall what all needs to be done
at this point.

To be continued...